merge@2019-06-30

+++ app/jobs/analysis_job.rb
@@ -4,18 +4,22 @@ class AnalysisJob < ActiveJob::Base
   def perform(analysis_id)
     analysis = Analysis.find(analysis_id)
     args = [analysis_id, analysis.num_data, analysis.num_tree]
-    output_dir = File.join(Rails.root, "/tmp/files/#{analysis_id}")
+    output_dir = Rails.root.join('tmp', 'files', analysis_id.to_s)
     FileUtils.mkdir_p(output_dir)
-    ret = system "Rscript #{Rails.root}/scripts/analyze.r #{args.join(' ')}"
+    success = system "Rscript #{Rails.root}/scripts/analyze.r #{args.join(' ')}"
+    raise StandardError unless success
 
     yaml_file = File.join(output_dir, 'analysis.yml')
-    if File.exists?(yaml_file)
+    if File.exist?(yaml_file)
       analysis_params = YAML.load_file(yaml_file)
-      analysis.update!(:num_feature => analysis_params['mtry'])
+      analysis.update!(num_feature: analysis_params['mtry'])
     end
 
-    analysis.update!(:state => 'completed')
-    AnalysisMailer.finished(analysis, ret).deliver_now
+    AnalysisMailer.completed(analysis).deliver_now
     FileUtils.rm_rf("#{Rails.root}/tmp/files/#{analysis_id}")
+    analysis.update!(state: 'completed')
+  rescue StandardError
+    analysis.update!(state: 'error')
+    AnalysisMailer.error(analysis).deliver_now
   end
 end
